# apps/miniapp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Bot Manager API",
    "description": "REST API behind the Telegram Mini App admin panel.\n\nAuthenticate by posting the WebApp `initData` string to `/api/auth/telegram`,\nthen send the returned token as `Authorization: Bearer <token>` on every other\nrequest. Tokens are short-lived; renew them with `/api/auth/refresh` rather than\nre-posting `initData`, which Telegram only lets you use once.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chats": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Chats where the caller is an administrator",
        "operationId": "listChats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatSummary"
                  },
                  "type": "array",
                  "title": "Response Listchats"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "TelegramSession": []
          }
        ]
      }
    },
    "/api/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "One chat with its module state and unlocked features",
        "operationId": "getChat",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetail"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "chats"
        ],
        "summary": "Update chat-level settings (language, timezone)",
        "operationId": "updateChat",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetail"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/admins/sync": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Re-read the chat's administrator list from Telegram",
        "description": "Refresh the mirror the chat list is drawn from, and drop cached verdicts.\n\nTelegram sends no update on a promotion, so this is how a newly promoted\nadmin makes the chat appear in their panel without waiting for the bot to\nhappen to sync it.",
        "operationId": "syncChatAdmins",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/bot-permissions": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Inspect the bot's live Telegram permissions for this chat",
        "operationId": "getChatBotPermissions",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatBotPermissions"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/plan": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "The plan actually in force for a chat",
        "operationId": "getChatPlan",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummary"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminRole": {
        "type": "string",
        "enum": [
          "owner",
          "admin"
        ],
        "title": "AdminRole"
      },
...
      "ChatBotPermissions": {
        "properties": {
          "reachable": {
            "type": "boolean",
            "title": "Reachable",
            "default": true
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "unknown"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          },
          "privacy_mode_disabled": {
            "type": "boolean",
            "title": "Privacy Mode Disabled",
            "default": false
          },
          "can_read_messages": {
            "type": "boolean",
            "title": "Can Read Messages",
            "default": false
          },
          "can_send_messages": {
            "type": "boolean",
            "title": "Can Send Messages",
            "default": false
          },
          "can_delete_messages": {
            "type": "boolean",
            "title": "Can Delete Messages",
            "default": false
          },
          "can_restrict_members": {
            "type": "boolean",
            "title": "Can Restrict Members",
            "default": false
          },
          "can_invite_users": {
            "type": "boolean",
            "title": "Can Invite Users",
            "default": false
          },
          "can_manage_topics": {
            "type": "boolean",
            "title": "Can Manage Topics",
            "default": false
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "title": "ChatBotPermissions"
      },
      "ChatDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tg_chat_id": {
            "type": "integer",
            "title": "Tg Chat Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "$ref": "#/components/schemas/ChatType"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "plan_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "role": {
            "$ref": "#/components/schemas/AdminRole",
            "default": "admin"
          },
          "members_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members Count"
          },
          "owner_tg_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Tg Id"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ru"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "modules": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Modules"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tg_chat_id",
          "title",
          "type",
          "plan"
        ],
        "title": "ChatDetail"
      },
      "ChatSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tg_chat_id": {
            "type": "integer",
            "title": "Tg Chat Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "$ref": "#/components/schemas/ChatType"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "plan_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "role": {
            "$ref": "#/components/schemas/AdminRole",
            "default": "admin"
          },
          "members_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tg_chat_id",
          "title",
          "type",
          "plan"
        ],
        "title": "ChatSummary"
      },
      "ChatType": {
        "type": "string",
        "enum": [
          "group",
          "supergroup",
          "channel"
        ],
        "title": "ChatType"
      },
      "ChatUpdate": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(ru|en)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "title": "ChatUpdate"
      },
...
      "OperationResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          }
        },
        "type": "object",
        "title": "OperationResult"
      },
...
      "Plan": {
        "type": "string",
        "enum": [
          "free",
          "pro",
          "business",
          "white_label"
        ],
        "title": "Plan"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "domain-error"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC7807-like error body returned by the API exception handler.\n\n`title` is localized for the caller and safe to show; `detail` is the\ndeveloper-facing message. `code` is the stable machine-readable discriminator\nthe Mini App branches on, and `context` carries whatever the error knew (the\nrequired plan for a paywall, the offending field for a validation failure)."
      },
...
    },
    "securitySchemes": {
      "TelegramSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}